dtdevdoc-8406-hosted-card-fields-temp-restore-v1-old
