+++ advanced-integration/v1/public/app.js
